board: move motors to B & C, move servo to H1
Servo move allows for the servo to be connected directly without any
intermediate jumpers.

Motor move allows for both motors to no longer share a H-bridge for
better heat dissipation. It also frees up PB3 so that we can output
trace data on that pin as well (just in case we need that functionality)