startup: enable ultrasonic sensor
No interurpt edge being selected caused the EXTI interrupt to
never fire.

TODO: actually enable the interrupt ONLY when the sensor is connected.
Enabling it now will cause suprious interrupts to be fired because of the
floating nature of the input pin. Since the input voltage to the pin is
5V, we cannot enable internal pull up / downs as that will cause the pin
to no longer be 5V tolerant.