trajectory_controller: always force steering movement before a move
The steering servo appears to have some kind of deadband where minor
deviations in output shaft angle are not corrected aggressively.

This causes issues when adjusting the servo angle when tuning the
left and right control signals - minor control signal changes lead
to unobservable changes in the output shaft angle because of that
deadband.

Force steering movement by having the first half of steering setup
time used to move the steering servo to the "opposite" of the sent
control signal.

The second half of the setup time is used to move the steering
servo to the "actual" commanded signal.

This avoids the deadband issue observed because we are always commanding
the servo to go from two states that are "far" from each other.